requirements.txt
streamlit==1.32.2
langchain==0.1.12
python-dotenv==1.0.1
uvicorn==0.29.0
fastapi==0.110.0
pydantic==2.6.4
google.generativeai==0.4.1
pandas==2.2.1
PyPDF2==3.0.1
faiss-cpu==1.8.0
langchain-google-genai==0.0.11